First iteration of the Reward System
Finished prototyping the reward system i.e. picks two random items, a stat reward and presents it to the player to choose. Logic for picking the reward is also implemented. Both the item and stat reward have a chance for 1 bonus stat reward for the player(can be positive or negative).